refactor identity config for tests and remove need to pass params to BrowserIDManager.
This patch moves the "mock" user and token code from test_browser_identity and
moves it into the test 'util' package, for use in subsequent patches (ie, so
other tests can also configure the BrowserIDManager for tests.)

util.js takes a 'config' object so the credentials can be configured by the
test - eg, some later tests want a specific username to be specified.  Later
patches allow this config object to also specify the basic credentials - so
tests can specify a single config object with required credentials for both
identity objects.  There is a getDefaultCredentials() method to get a default
config object - this will be used if no config is passed, or tests can grab
the default and override only the attributes they care about (such as username,
which later patches do.)

This patch also makes BrowserIDManager take no params to its constructor - the
test code just replaces the this._fxaService and this._tokenServerClient with
the mock versions - this keeps some of the other code cleaner - code that
creates a BrowserIDManager no longer need to import additional modules and
create objects just for the ctor.